PYTHONCODES/pyproject.toml
[project]
name = "pythoncodes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "pyperclip>=1.11.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "twilio>=9.10.1",
]
